Restore reconciliation functionality
Change structure and naming a little bit:
- Expose interface and trigger access on Astarte.Interface and Astarte.Trigger,
resembling how Ash resources work. This are not currently Ash resources but they
could become them in the future.
- Introduce the DataLayer concept to express the actual implementation of the
mechanism to retrieve data. This is the same mechanism there was in place
before, it just has a precise name now.
