+++ backend/lib/edgehog/application.ex
@@ -53,12 +53,11 @@ defmodule Edgehog.Application do
       {Finch, name: EdgehogFinch},
       # Start the UpdateCampaigns supervisor
       Edgehog.UpdateCampaigns.Supervisor,
+      # Start the Tenant Reconciler Supervisor
+      {Edgehog.Tenants.Reconciler.Supervisor,
+       tenant_to_trigger_url_fun: tenant_to_trigger_url_fun},
       # Start the Endpoint (http/https)
       EdgehogWeb.Endpoint
-      # TODO: enable this back after Ash porting is complete
-      # # Start the Tenant Reconciler Supervisor
-      # {Edgehog.Tenants.Reconciler.Supervisor,
-      #  tenant_to_trigger_url_fun: tenant_to_trigger_url_fun}
     ]
 
     # See https://hexdocs.pm/elixir/Supervisor.html
