+++ backend/lib/edgehog/tenants/reconciler/core.ex
@@ -20,7 +20,8 @@
 
 defmodule Edgehog.Tenants.Reconciler.Core do
   alias Astarte.Client
-  alias Edgehog.Astarte
+  alias Edgehog.Astarte.Interface
+  alias Edgehog.Astarte.Trigger
   alias Edgehog.Tenants.Reconciler.AstarteResources
 
   @interfaces AstarteResources.load_interfaces()
@@ -45,7 +46,7 @@ defmodule Edgehog.Tenants.Reconciler.Core do
       "version_minor" => required_minor
     } = required_interface
 
-    case Astarte.fetch_interface(client, interface_name, required_major) do
+    case Interface.fetch_by_name_and_major(client, interface_name, required_major) do
       {:ok, %{"version_major" => ^required_major, "version_minor" => existing_minor}} ->
         if required_minor > existing_minor do
           update_interface!(client, interface_name, required_major, required_interface)
@@ -66,7 +67,7 @@ defmodule Edgehog.Tenants.Reconciler.Core do
       "name" => trigger_name
     } = required_trigger
 
-    case Astarte.fetch_trigger(client, trigger_name) do
+    case Trigger.fetch_by_name(client, trigger_name) do
       {:ok, existing_trigger} ->
         if existing_trigger != required_trigger do
           update_trigger!(client, trigger_name, required_trigger)
@@ -84,14 +85,14 @@ defmodule Edgehog.Tenants.Reconciler.Core do
 
   defp update_interface!(rm_client, name, major, interface_json) do
     # TODO: crash with a nicer exception instead of MatchError
-    :ok = Astarte.update_interface(rm_client, name, major, interface_json)
+    :ok = Interface.update(rm_client, name, major, interface_json)
 
     :ok
   end
 
   defp install_interface!(rm_client, interface_json) do
     # TODO: crash with a nicer exception instead of MatchError
-    :ok = Astarte.create_interface(rm_client, interface_json)
+    :ok = Interface.create(rm_client, interface_json)
 
     :ok
   end
@@ -100,7 +101,7 @@ defmodule Edgehog.Tenants.Reconciler.Core do
     # Triggers don't have a way to update them, to we delete it and install it again
 
     # TODO: crash with a nicer exception instead of MatchError
-    :ok = Astarte.delete_trigger(rm_client, name)
+    :ok = Trigger.delete(rm_client, name)
     :ok = install_trigger!(rm_client, trigger_json)
 
     :ok
@@ -108,7 +109,7 @@ defmodule Edgehog.Tenants.Reconciler.Core do
 
   defp install_trigger!(rm_client, trigger_json) do
     # TODO: crash with a nicer exception instead of MatchError
-    :ok = Astarte.create_trigger(rm_client, trigger_json)
+    :ok = Trigger.create(rm_client, trigger_json)
 
     :ok
   end
